requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
sqlite3